[ParameterizedGenerator] Add resource size and count to trace params
This is useful for testing the impact of "shard-by-trace-id" architectures.

Previously, all exported batches only included a single trace, which means they
always end up in the same shard. Now it's possible to touch multiple shards per
export request, which in some cases is important to properly demonstrate the
downstream system's performance.